Changes made to this commit: renamed:    src/loading.json -> src/assets/loading.json renamed:    src/notfound.json -> src/assets/notfound.json 	Moved animation files into assets modified:   src/components/App.tsx 	For hero name heading (on hover) I changed it to useRef instead 	of useState because it's not a good idea having the App 	component to re-render everytime we hover on a hero. useRef is 	like a variable storing an element which later we can access it 	and change its textContent to hovering hero's name. modified:   src/components/FilterOption.tsx 	Change prop name names to heroNames. modified:   src/components/HeroCard.tsx 	Removed variables an eject expressions right into the render 	statement, also changed from ternary to condition && value when 	I wanted to return only one value. modified:   src/components/HeroDetail.tsx modified:   src/components/NotFound.tsx 	Changed animation import path.
