+++ src/components/FilterOption.tsx
@@ -5,7 +5,7 @@ type FilterOptionProps = {
   filterAttackType: (e: ChangeEvent<HTMLSelectElement>) => void,
   filterName: (e: ChangeEvent<HTMLSelectElement>) => void,
   clearFilter: () => void,
-  names: Array<string>,
+  heroNames: Array<string>,
   currentRole: string,
   currentAttackType: string,
   currentName: string,
@@ -13,7 +13,7 @@ type FilterOptionProps = {
 
 // If you have a type applied you can extract properties from the props.
 export const FilterOption: React.FC<FilterOptionProps> = ({
-  filterRole, filterAttackType, filterName, clearFilter, names, currentRole, currentAttackType, currentName
+  filterRole, filterAttackType, filterName, clearFilter, heroNames, currentRole, currentAttackType, currentName
 }): JSX.Element => {
 
   const roles: Array<string> = [
@@ -37,8 +37,7 @@ export const FilterOption: React.FC<FilterOptionProps> = ({
 
   const roleOptions = generateOptions(roles, currentRole)
   const attackTypeOptions = generateOptions(attackTypes, currentAttackType)
-  const nameOptions = generateOptions(names, currentName)
-
+  const nameOptions = generateOptions(heroNames, currentName) 
   return (
     <div className="filter">
       <p>FILTER</p>
