+++ src/components/HeroCard.tsx
@@ -16,19 +16,17 @@ type HeroCardProps = {
 export const HeroCard: React.FC<HeroCardProps> = ({
   id, mouseOver, attr, localized_name, name, isValid
 }): JSX.Element => {
-  const derivedClass = `${isValid && "hover"} ${isValid && attr}`
-  const derivedOpacity = isValid ? 1 : .3
-  const derivedMouseOver = isValid ? mouseOver  : (): void => {}
   return (
     <Link to={`/${localized_name}`}>
       <div className="hero-card">
         <img
           key={`hero-${id}`}
-          onMouseOver={derivedMouseOver}
-          className={derivedClass}
+          // It's like ternary but only return one value if the condition is true
+          onMouseOver={isValid ? mouseOver : () => {}}
+          className={`${isValid && "hover"} ${isValid && attr}`}
           data-id={id}
           src={`https://cdn.dota2.com/apps/dota2/images/heroes/${name}_full.png`}
-          style={{opacity: derivedOpacity}}
+          style={{opacity: isValid ? 1 : .3}}
           alt={name}
         />  
       </div>
